Remove uncertainty about copyright owner
I am making an experimental SVG 3D animation but I am uncertain who/how to attribute the code I am using from here and want to include the license but the license is missing the copyright year and owner. I am assuming it is Jake Archibald and set the year to the current year.